% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{baseball}
\alias{baseball}
\title{1.1 List}
\usage{
baseball(text = "a <- 1+1", keep.source = TRUE)
}
\arguments{
\item{text}{source code of R. Required.}
}
\value{
ggplot object.
}
\description{
Returns a row with the latest price for a given stock.
}
