% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rpcTest.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(myname = "")
}
\arguments{
\item{myname}{your name. Required.}
}
\description{
Basic hello world function to be called from the demo app
}

