% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/scan.R
\docType{data}
\name{json_file}
\alias{json_file}
\title{scan}
\format{\preformatted{ chr "http://webonastick.com/uscl/feeds/uscl.json.txt"
}}
\usage{
json_file
}
\arguments{
\item{type}{type of plot to create. One of "smoothplot", "highlowplot", "areaplot".}

\item{ticker}{stock ticker symbol. E.g. "GOOG".}

\item{from}{start date. Either string or date object.}

\item{to}{end date. Either string or date object.}

\item{current}{include the current price of this stock. TRUE/FALSE.}
}
\description{
This is a wrapper for the OpenCPU application. It is a single function that calls out to various plot types.
The function prints a plot to the graphics device and returns nothing.
}
\keyword{datasets}

