% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/scan.R
\name{scanR}
\alias{scanR}
\title{scan}
\usage{
scanR()
}
\description{
This is a wrapper for the OpenCPU application. It is a single function that calls out to various plot types.
The function prints a plot to the graphics device and returns nothing.
}

