% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{server}
\alias{server}
\title{1.2 socket}
\usage{
server()
}
\arguments{
\item{text}{source code of R. Required.}
}
\value{
ggplot object.
}
\description{
Returns a row with the latest price for a given stock.
}
