% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/testR.R
\name{test}
\alias{test}
\title{Test}
\usage{
test(ticker = "GOOG")
}
\arguments{
\item{ticker}{stock ticker symbol. E.g. "GOOG".}
}
\description{
Returns a row with the latest price for a given stock.
}

